move http adapter into http-adapter folder
